Mysteriously fix hot reloading
I'm not sure why hot reloading wasn't working or why this fixes it, but it
seems to work now. The only clue I got was from the Webpack Development
Guide (https://webpack.js.org/guides/development/):

"Warning: webpack-dev-server doesn't write any output files after
compiling. Instead, it keeps bundle files in memory and serves them as if
they were real files mounted at the server's root path. If your page
expects to find the bundle files on a different path, you can change this
with the devMiddleware.publicPath option in the dev server's
configuration."

So, I looked under the devMiddleware config options and tried writeToDisk,
and at that point the app started hot-reloading.

I'm not sure if this is the correct solution though, because it also
writes the hot-update code to the bundle that is published to the web.