Keep hot-reloading out of the published bundle
If we run the dev server for development but then make sure to run
`webpack build` before committing, hot-reloading will work but won't show
up on the published web version.

Add package.json scripts 'dev' and 'build' to do this, plus a reminder in
the README. Also remove the hot-reload files from gitignore so that it's
clear if we forget to run build before committing.