Implement some Math functions
In order to make Math functions work with Decimals, we replace them in
patches.js with versions that handle both Decimal and non-Decimal
arguments.

Most Math methods have only one argument so the override is short, but for
methods such as Math.pow() with two or more arguments we also have to
check for mixed-type arguments and throw a TypeError.

This adds tests for how the overrides behave at runtime. In order to make
the tests work, we have to add the decimal.js and big.js libraries to
package.json with the exact versions pinned that we load in
src/runner/index.html. Otherwise it's confusing, since there were other
dependencies that required earlier versions of big.js with less
functionality.