Implement Decimal.round()
Adds implementation to the patches file, and some logic in the transforms
to mark the return value node as known-decimal.

There are conflicting definitions in the proposal repo about the API of
Decimal.round: whether the options are required or not. I've gone with the
definition in the README for now.