Make 'unimplemented' error message consistent
We now have two placeholders for functionality that isn't implemented yet
because big.js doesn't support it out of the box. Let's have a utility
function that throws a consistent error message that will hopefully point
people to the proposal repo to give us feedback if they need something
that's missing.