Consolidate lists of patched Math methods
Now that the patches file is a module, we can remove this "keep in sync"
comment that I added earlier, and just ensure programmatically that the
Math methods that the transform knows about are all patched and vice
versa.

If we forget to update the list, then we should get an early warning via
the tests, instead of the much-harder-to-debug error where the output of
the Math method won't be marked as a Decimal node.