Fix unary - for BigDecimal
The .neg() method doesn't actually exist on big.js instances. To negate a
big.js instance, we have to use .mul(-1).

This removes replaceWithUnaryDecimalExpression out of the shared transform
code and into the specific transforms, since we can no longer perform the
same transformation for Decimal128 and BigDecimal, or easily express the
differences in a table of method names like we do with `sharedOpts` for
the binary expressions.

Closes: #26