Add examples dropdown menu
This moves the introductory text to its own "Intro" example which is
pre-loaded the first time you visit the playground.

It adds a dropdown menu to the top right, where you can select an example
to be loaded. The state of the dropdown menu is not saved in the hash.

Since we need to change the contents of the code editor programmatically
now, this passes the code from App into the Editor render function as a
parameter, rather than getting it from the editor model. The editor model
is now unused and doesn't need to be created explicitly to pass into the
editor component. (I'm not sure about this though. When I tried updating
the state of the editor model, it didn't change what was rendered, but
maybe I was doing it wrong.)

In order to have something else to select in the dropdown menu, this also
adds a second example, "Scientific notation", which demonstrates
formatting a Decimal quantity using scientific notation. (Note, this
example requires Math.log10 so it doesn't yet work in BigDecimal.)

The examples are stored as long strings in src/examples.js.