Be more explicit about what goes on the global object in globals.js
Anything defined at the top level in this file is somewhat unobviously
added to the global object. Make this more explicit, and prevent pollution
by names that don't need to be there, by changing globals.js to be a
module which exports a defineGlobals() function, and calling that function
in the HTML preamble.

This allows deduplicating the SHARED_SINGLE_OPS and SHARED_MIXED_OPS
constants from constants.js, but SHARED_SINGLE_OPS is actually not the
same one as in that file, so we need to clone and modify it.