Add initial examples
This adds Sarah's two examples from
https://github.com/sarahghp/decimal-playground/pull/47 into the examples
pane, with syntax highlighting, from where they can be copy-pasted into
the code editor.

The introductory code sample becomes the "Adding with Logging" example,
which is pre-loaded the first time you visit the playground, so that
first-time users know about the quirks of using console.log().

The DOM manipulation moves to its own example. This example is a
placeholder because it doesn't actually use Decimals. The description has
a note to that effect. We plan to replace it later with a calculator app.

The examples are stored as strings, in files inside src/examples/ and
consolidated as one import in src/examples/index.js. They are linted as
part of the test suite and during 'npm run lint'.

Co-authored-by: Sarah GHP <sarah.groff.palermo@gmail.com>