Add button for copying examples to editor
Adds a button floating over the top right corner of the code example,
which copies it to the editor. The button is normally translucent so as
not to overlap the code, but becomes solid on hover. As user feedback when
clicked, the button turns a shade darker, and displays "Copied!" for 2
seconds.

Based on Sarah's proof of concept:
https://github.com/sarahghp/decimal-playground/pull/47

Co-authored-by: Sarah GHP <sarah.groff.palermo@gmail.com>